+++ src/containers/digest/CommunityDigest/ClassicLayout/CommunityBrief.tsx
@@ -3,9 +3,7 @@ import { contains } from 'ramda'
 
 import type { TCommunity } from '@/spec'
 import { ICON_CMD } from '@/config'
-import { HCN, NON_FILL_COMMUNITY } from '@/constant'
-
-import CommunityJoinSign from '@/widgets/CommunityJoinSign'
+import { NON_FILL_COMMUNITY } from '@/constant'
 
 import {
   Wrapper,
@@ -17,7 +15,7 @@ import {
   TitleText,
   LogoHolder,
 } from '../styles/classic_layout/community_brief'
-import { subscribeCommunity, unsubscribeCommunity } from '../logic'
+// import { subscribeCommunity, unsubscribeCommunity } from '../logic'
 
 const CommunityLogoHolder = `${ICON_CMD}/community_logo_holder.svg`
 
@@ -43,13 +41,6 @@ const CommunityBrief: FC<TProps> = ({ community }) => {
         <TitleWrapper>
           <Title>
             <TitleText>{community.title}</TitleText>
-            {community.raw !== HCN && (
-              <CommunityJoinSign
-                onFollow={() => subscribeCommunity(community.id)}
-                onUndoFollow={() => unsubscribeCommunity(community.id)}
-                hasFollowed={community.viewerHasSubscribed}
-              />
-            )}
           </Title>
         </TitleWrapper>
       </CommunityInfo>
